+++ integration-tests/karura-runtime-mock/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "runtime-integration-tests"
-version = "0.7.4"
-description = "Integration tests"
+name = "karura-runtime-mock"
+version = "0.1.0"
+description = "A mock runtime for Karura parachain"
 authors = ["GalacticCouncil"]
 edition = "2021"
 homepage = "https://github.com/galacticcouncil/Basilisk-node"
@@ -18,15 +18,15 @@ pallet-price-oracle = { git = "https://github.com/galacticcouncil/warehouse", re
 pallet-nft = { git = "https://github.com/galacticcouncil/warehouse", rev = "85257f53b0da5827c103bf4f739474cc1e83d3f7", default-features = false }
 hydradx-adapters = { git = "https://github.com/galacticcouncil/warehouse",rev = "85257f53b0da5827c103bf4f739474cc1e83d3f7", default-features = false }
 
-pallet-exchange = { path = "../pallets/exchange", default-features=false}
-pallet-exchange-benchmarking = { path = "../pallets/exchange/benchmarking", optional = true, default-features = false}
-pallet-xyk = { path = "../pallets/xyk",default-features = false}
-pallet-duster= { path = "../pallets/duster",default-features = false}
-pallet-xyk-rpc-runtime-api = { path = "../pallets/xyk/rpc/runtime-api",default-features = false}
-pallet-lbp = { path = "../pallets/lbp", default-features = false }
-pallet-marketplace = { path = '../pallets/marketplace', default-features = false }
+pallet-exchange = { path = "../../pallets/exchange", default-features=false}
+pallet-exchange-benchmarking = { path = "../../pallets/exchange/benchmarking", optional = true, default-features = false}
+pallet-xyk = { path = "../../pallets/xyk",default-features = false}
+pallet-duster= { path = "../../pallets/duster",default-features = false}
+pallet-xyk-rpc-runtime-api = { path = "../../pallets/xyk/rpc/runtime-api",default-features = false}
+pallet-lbp = { path = "../../pallets/lbp", default-features = false }
+pallet-marketplace = { path = '../../pallets/marketplace', default-features = false }
 
-primitives = { default-features = false, path = "../primitives" }
+primitives = { default-features = false, path = "../../primitives" }
 
 pallet-treasury = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 pallet-democracy = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
@@ -85,7 +85,6 @@ pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "
 pallet-transaction-payment = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 pallet-transaction-payment-rpc-runtime-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 pallet-utility = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-pallet-uniques = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 sp-block-builder = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 sp-consensus-aura = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
@@ -101,24 +100,26 @@ sp-version = { git = "https://github.com/paritytech/substrate", branch = "polkad
 sp-staking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 sp-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
 sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
+sp-arithmetic = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17" }
+
+xcm-emulator = { git = "https://github.com/zqhxuyuan/xcm-simulator", branch = "v0.9.16" }
+
 
 polkadot-primitives = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.17" }
 polkadot-runtime-parachains = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.17" }
 
-basilisk-runtime = { path = "../runtime/basilisk", default-features = false}
+basilisk-runtime = { path = "../../runtime/basilisk", default-features = false}
 kusama-runtime = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.17" }
-karura-runtime-mock = { path = "karura-runtime-mock", default-features = false}
 
 [dev-dependencies]
 hex-literal = "0.3.1"
 pallet-relaychain-info = { git = "https://github.com/galacticcouncil/warehouse", rev = "85257f53b0da5827c103bf4f739474cc1e83d3f7" }
 pretty_assertions = "1.2.1"
-xcm-emulator = { git = "https://github.com/zqhxuyuan/xcm-simulator", branch = "v0.9.16" }
-sp-arithmetic = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17" }
-test-case = "2.2.1"
+env_logger = "*"
 
 [features]
 default = ["std"]
+no_std = []
 std = [
     "frame-executive/std",
     "frame-support/std",
